junzijunshop/shop/requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.7.0
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2016.2.28
colorama==0.4.1
decorator==4.4.0
Django==2.1.8
django-celery==3.3.1
django-ckeditor==5.7.1
django-filter==2.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-redis-cache==2.0.0
djangorestframework==3.10.3
flower==0.9.3
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
kombu==3.0.37
Markdown==3.1.1
mypython==1.0.0
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
pycryptodomex==3.7.2
Pygments==2.4.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.2
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
wincertstore==0.2
